removed default config function
